Refactored stuff into dashboard component and did code cleanup.
